Move #include utils to common folder
